Raise the toppiness
The clock was being covered by notifications. That's annoying when a
Teams meeting start notification pops up and I can't see how long
until the meeting's actually supposed to start.

This level puts it over notifications and also the menu bar. Doesn't
seem like there's a setting between those. This means the clock will
cover whatever part of the menu bar it's over. This is not a huge deal
when it's positioned right under where the clock normally goes but
probably isn't so great in general.

In the future this should be a settable option.